requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.0
stripe==9.12.0
typing-extensions==4.12.2
urllib3==2.2.2
